Generalized HLSL shader compilation
