+++ src/core/hle/D3D8/Direct3D9/VertexShader.h
@@ -1,6 +1,4 @@
-
-#ifndef DIRECT3D9VERTEXSHADER_H
-#define DIRECT3D9VERTEXSHADER_H
+#pragma once
 
 #include "core\hle\D3D8\XbVertexShader.h"
 #include "FixedFunctionVertexShaderState.hlsli"
@@ -15,7 +13,7 @@ static const char* vs_model_2_a = "vs_2_a";
 static const char* vs_model_3_0 = "vs_3_0";
 extern const char* g_vs_model;
 
-extern HRESULT EmuCompileShader
+extern HRESULT EmuCompileVertexShader
 (
     IntermediateVertexShader* pIntermediateShader,
     ID3DBlob** ppHostShader
@@ -25,4 +23,3 @@ extern void EmuCompileFixedFunction(ID3DBlob** ppHostShader);
 
 extern HRESULT EmuCompileXboxPassthrough(ID3DBlob** ppHostShader);
 
-#endif
